requirements.txt
fastapi
uvicorn
openai
google-generativeai
groq
requests
python-dotenv
streamlit
